+++ src/renderer/features/local-conversation/view/composer/local-conversation-thread-composer.stories.tsx
@@ -59,7 +59,7 @@ function buildActions(): ThreadStageActions {
     onRespondApproval: async () => { },
     onRespondUserInput: async () => { },
     onRespondMcpElicitation: async () => { },
-    onResolvePlanImplementationRequest: () => { },
+    onResolvePlanImplementationRequest: async () => { },
     onEnqueueQueuedFollowUp: async () => { },
     onRemoveQueuedFollowUp: async () => { },
     onReorderQueuedFollowUps: async () => { },